requirements.txt
python-telegram-bot[job-queue]==20.8
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic==2.7.4
pydantic-settings==2.3.4
httpx==0.27.0
python-dotenv==1.0.1
aiohttp==3.9.5
apscheduler==3.10.4
Flask==3.0.3
